app/requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
pandas==1.3.5
pymoo==0.6.1.1
nltk==3.7
newsapi-python==0.2.7
yfinance==0.2.40
mangum==0.17.0
numpy==1.23.5